Extract view plugin
