+++ main.ts
@@ -14,15 +14,7 @@ import {
 import { debounce, Debouncer, MarkdownView, Plugin, setIcon } from "obsidian";
 import { DEFAULT_SETTINGS, PluginSettings } from "./settings";
 import { SimpleEmbedPluginSettingTab } from "./settings-tab";
-import { RangeSetBuilder } from "@codemirror/rangeset";
-import {
-  Decoration,
-  DecorationSet,
-  EditorView,
-  ViewPlugin,
-  ViewUpdate,
-  WidgetType,
-} from "@codemirror/view";
+import { buildSimpleEmbedsViewPlugin } from "./view-plugin";
 
 export default class SimpleEmbedsPlugin extends Plugin {
   settings: PluginSettings;
@@ -48,7 +40,7 @@ export default class SimpleEmbedsPlugin extends Plugin {
 
     this.currentTheme = this._getCurrentTheme();
 
-    const ext = this.buildSimpleEmbedsViewPlugin(this);
+    const ext = buildSimpleEmbedsViewPlugin(this);
     this.registerEditorExtension(ext);
 
     this.processedMarkdown = debounce(() => {
@@ -142,7 +134,7 @@ export default class SimpleEmbedsPlugin extends Plugin {
     }
   }
 
-  private createEmbed(
+  createEmbed(
     embedSource: EmbedSource,
     link: string,
     fullWidth: boolean,
@@ -180,125 +172,5 @@ export default class SimpleEmbedsPlugin extends Plugin {
   get isLivePreviewSupported(): boolean {
     return (this.app.vault as any).config?.livePreview &&
       this.settings.enableInLivePreview;
-  }
-
-  buildSimpleEmbedsViewPlugin(plugin: SimpleEmbedsPlugin) {
-    class EmbedWidget extends WidgetType {
-      constructor(
-        readonly link: string,
-        readonly embedSource: EmbedSource,
-        readonly plugin: SimpleEmbedsPlugin,
-      ) {
-        super();
-      }
-
-      eq(other: EmbedWidget) {
-        return other.link === this.link;
-      }
-
-      toDOM() {
-        const embed = this.plugin.createEmbed(
-          this.embedSource,
-          this.link,
-          false,
-        );
-        return embed;
-      }
-
-      ignoreEvent() {
-        return true;
-      }
-    }
-
-    const viewPlugin = ViewPlugin.fromClass(
-      class {
-        decorations: DecorationSet;
-
-        constructor(view: EditorView) {
-          this.decorations = this.buildDecorations(view);
-        }
-
-        update(update: ViewUpdate) {
-          if (
-            update.docChanged || update.viewportChanged || update.selectionSet
-          ) {
-            this.decorations = this.buildDecorations(update.view);
-          }
-        }
-
-        destroy() {
-        }
-
-        buildDecorations(view: EditorView) {
-          let builder = new RangeSetBuilder<Decoration>();
-
-          if (!plugin.isLivePreviewSupported) {
-            return builder.finish();
-          }
-
-          let lines: number[] = [];
-          if (view.state.doc.length > 0) {
-            lines = Array.from(
-              { length: view.state.doc.lines },
-              (_, i) => i + 1,
-            );
-          }
-
-          const currentSelections = [...view.state.selection.ranges];
-
-          for (let n of lines) {
-            const line = view.state.doc.line(n);
-            const startOfLine = line.from;
-            const endOfLine = line.to;
-
-            let currentLine = false;
-
-            currentSelections.forEach((r) => {
-              if (r.from >= startOfLine && r.to <= endOfLine) {
-                currentLine = true;
-                return;
-              }
-            });
-
-            const mdLink = line.text.match(/\[.*\]\(\S*\)/)?.first().trim();
-            if (!currentLine && mdLink) {
-              const start = line.text.indexOf(mdLink) + startOfLine;
-              const end = start + mdLink.length;
-              let embedSource = plugin.embedSources.find((source) => {
-                return plugin.settings[source.enabledKey] &&
-                  source.regex.test(line.text);
-              });
-              if (embedSource) {
-                const link = line.text.match(embedSource.regex).first();
-                const deco = Decoration.replace({
-                  widget: new EmbedWidget(
-                    link,
-                    embedSource,
-                    plugin,
-                  ),
-                  inclusive: false,
-                });
-                if (plugin.settings.keepLinksInPreview) {
-                  if (plugin.settings.embedPlacement === "above") {
-                    builder.add(start, start, deco);
-                  } else if (plugin.settings.embedPlacement === "below") {
-                    builder.add(end, end, deco);
-                  }
-                } else {
-                  builder.add(start, end, deco);
-                }
-              }
-            }
-          }
-
-          return builder.finish();
-        }
-      },
-      {
-        decorations: (v) => v.decorations,
-      },
-    );
-
-    return viewPlugin;
   }
 }
